shift full uint256

+++ src/security-council-mgmt/governors/modules/SecurityCouncilMemberElectionGovernorCountingUpgradeable.sol
@@ -203,7 +203,7 @@ abstract contract SecurityCouncilMemberElectionGovernorCountingUpgradeable is
         uint256[] memory topNomineesPacked = new uint256[](k);
 
         for (uint16 i = 0; i < nominees.length; i++) {
-            uint240 weight = weights[i];
+            uint256 weight = weights[i];
             uint256 packed = (weight << 16) | i;
 
             if (topNomineesPacked[0] < packed) {
